+++ Sora/Signaling.swift
@@ -330,7 +330,12 @@ public struct SignalingConnect {
     public var redirect: Bool?
 
     /// 転送フィルターの設定
+    ///
+    /// この項目は 2025 年 12 月リリース予定の Sora にて廃止されます
     public var forwardingFilter: ForwardingFilter?
+
+    /// リスト形式の転送フィルターの設定
+    public var forwardingFilters: [ForwardingFilter]?
 
     /// VP9 向け映像コーデックパラメーター
     public var vp9Params: Encodable?
@@ -784,6 +789,7 @@ extension SignalingConnect: Codable {
         case audio_streaming_language_code
         case redirect
         case forwarding_filter
+        case forwarding_filters
     }
 
     enum VideoCodingKeys: String, CodingKey {
@@ -834,6 +840,7 @@ extension SignalingConnect: Codable {
         try container.encodeIfPresent(audioStreamingLanguageCode, forKey: .audio_streaming_language_code)
         try container.encodeIfPresent(redirect, forKey: .redirect)
         try container.encodeIfPresent(forwardingFilter, forKey: .forwarding_filter)
+        try container.encodeIfPresent(forwardingFilters, forKey: .forwarding_filters)
 
         if videoEnabled {
             if videoCodec != .default || videoBitRate != nil || vp9Params != nil || av1Params != nil || h264Params != nil {