+++ CHANGES.md
@@ -20,6 +20,11 @@
   - libwebrtc の対象バージョンに追従した
     - <https://webrtc.googlesource.com/src/+/9b81d2c954128831c62d8a0657c7f955b3c02d32>
   - @miosakuma
+- [ADD] `ForwardingFilter` に name と priority を追加する
+  - @zztkm
+- [ADD] シグナリング connect 時にリスト形式の転送フィルターを設定するための項目を追加する
+  - `Configuration`, `SignalingConnect` に forwardingFilters を追加する
+  - @zztkm
 
 ### misc